\name{ctrldefault}
\alias{ctrldefault}
\alias{dispdefault}
\encoding{UTF-8}
\title{set control defaults}
\concept{minimization}
\concept{maximization}
\description{
  Set control defaults.
}
\usage{
   ctrldefault(npar)

   dispdefault(ctrl)
}
\arguments{
 \item{npar}{Number of parameters to optimize.}
 \item{ctrl}{A list (likely generated by `ctrldefault`) of default settings to `optimx`.}
}
\value{
  \code{ctrldefault} returns the default control settings for optimization tools.

  \code{dispdefault} provides a compact display of the contents of a control settings list.
}
\keyword{nonlinear}
\keyword{optimize}

